// src/config/enums.ts
export enum RedisFailureStrategy {
  FAIL_OPEN = 'fail-open',
  FAIL_CLOSED = 'fail-closed',
}

export enum ThrottlerErrorCode {
  INVALID_CONFIGURATION = 'INVALID_CONFIGURATION',
  MISSING_REQUIRED_CONFIG = 'MISSING_REQUIRED_CONFIG',
  REDIS_CONNECTION_FAILED = 'REDIS_CONNECTION_FAILED',
  REDIS_OPERATION_FAILED = 'REDIS_OPERATION_FAILED',
  REDIS_FUNCTIONS_LOAD_FAILED = 'REDIS_FUNCTIONS_LOAD_FAILED',
  STORAGE_OPERATION_FAILED = 'STORAGE_OPERATION_FAILED',
  UNKNOWN_ERROR = 'UNKNOWN_ERROR',
}